Fix minor bugs and codestyle
